Extract server functions from main.rs to lib.rs
